# api/openapi.yaml
openapi: 3.0.3
info:
  title: FQP From Scratch API
  version: 1.0.0
  description: 官方足彩/竞彩足球量化研究与长期运行系统 API 草案。系统只提供数据分析、模拟与复盘接口，不提供购彩、代购、出票、收款接口。
servers:
  - url: http://localhost:8000
paths:
...
  /api/v1/tickets/ocr:
    post:
      summary: 上传实票照片并进行 OCR 识别
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200': { description: OCR 识别结果，包含 ticket_image_url 原图地址 }
  /api/v1/real-tickets/upload:
    post:
      summary: 上传线下自购彩票照片并进行 OCR 识别
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                purchase_time:
                  type: string
                total_amount:
                  type: number
      responses:
        '200': { description: OCR 识别结果，确认后通过 /api/v1/real-tickets 创建实票记录 }
  /api/v1/real-tickets:
    get:
      summary: 查询实票列表
      parameters:
        - name: status
          in: query
          required: false
          schema: { type: string }
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 20 }
      responses:
        '200': { description: 实票列表 }
    post:
      summary: 手动创建已确认的线下自购实票记录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ticket, items]
              properties:
                ticket:
                  type: object
                items:
                  type: array
                  items: { type: object }
      responses:
        '200': { description: 实票创建结果 }
  /api/v1/real-tickets/{id}/confirm:
    post:
      summary: 用户确认/校对票据信息
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: 已确认 }
  /api/v1/real-tickets/{ticket_id}:
    get:
      summary: 查询单张实票及投注项
      parameters:
        - name: ticket_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: 实票详情 }
    put:
      summary: 更新实票校对信息、状态或绑定关系
      parameters:
        - name: ticket_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: 更新结果 }
    delete:
      summary: 删除实票记录
      parameters:
        - name: ticket_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: 删除结果 }
  /api/v1/real-tickets/{id}/settle:
    post:
      summary: 触发单张实票结算
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: 结算结果 }
  /api/v1/reviews/daily:
    get:
      summary: 查询日报
      parameters:
        - name: date
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        '200': { description: 日复盘 }
  /api/v1/reviews/weekly:
    get:
      summary: 查询周复盘
      parameters:
        - name: week_start
          in: query
          required: false
          schema: { type: string, format: date }
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 12 }
      responses:
        '200': { description: 周复盘列表 }
  /api/v1/reviews/monthly:
    get:
      summary: 查询月复盘
      parameters:
        - name: month
          in: query
          required: false
          schema: { type: string, pattern: '^\d{4}-\d{2}$' }
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 12 }
      responses:
        '200': { description: 月复盘列表 }
...
  /api/v1/modules/{module_code}/status:
    patch:
      summary: 更新运行时模块启停状态
      tags: [ui]
      parameters:
        - name: module_code
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [disabled]
              properties:
                disabled: { type: boolean }
      responses:
        '200': { description: 模块状态已更新 }
        '404': { description: 模块不存在 }
        '409': { description: 必需模块或依赖关系阻止更新 }
  /api/v1/ui/panels:
    get:
      summary: 查询可见功能面板
      tags: [ui]
      parameters:
        - name: disabledModules
          in: query
          required: false
          schema:
            type: array
            items: { type: string }
      responses:
        '200': { description: 按功能顺序返回的面板注册表 }